Move to typescript and divde into more exports
