+++ package.json
@@ -6,7 +6,8 @@
   "type": "module",
   "scripts": {
     "prepare": "husky install",
-    "prettier": "prettier --write ."
+    "prettier": "prettier --write .",
+    "build": "tsc -b"
   },
   "keywords": [],
   "author": "",
@@ -19,6 +20,8 @@
     "@commitlint/cli": "^17.0.3",
     "@commitlint/config-conventional": "^17.0.3",
     "@release-it/conventional-changelog": "^5.0.0",
+    "@types/inquirer": "^9.0.0",
+    "@types/node": "^18.6.4",
     "husky": "^8.0.1",
     "prettier": "^2.7.1",
     "release-it": "^15.2.0"
